requirements.txt
typer[all]
hatch
pytest
PyGithub
requests